+++ test/planck.js
@@ -3,7 +3,7 @@ import * as namespace from '../index';
 
 const should = chai.should();
 
-describe('planck entry point should be namespace with all main classes:', () => {
+describe('planck entry point should be namespace with all main classes and namespaces:', () => {
 	it('App', () => {
 		should.exist(namespace.App);
 		namespace.App.should.be.a('function');
@@ -28,6 +28,32 @@ describe('planck entry point should be namespace with all main classes:', () => 
 	it('Controller', () => {
 		should.exist(namespace.Controller);
 		should.exist(namespace.Controller.Base);
+		should.exist(namespace.Controller.useViews);
+		should.exist(namespace.Controller.before);
+		should.exist(namespace.Controller.skipBefore);
+		should.exist(namespace.Controller.after);
+		should.exist(namespace.Controller.skipAfter);
 		namespace.Controller.Base.should.be.a('function');
+		namespace.Controller.useViews.should.be.a('function');
+		namespace.Controller.before.should.be.a('function');
+		namespace.Controller.skipBefore.should.be.a('function');
+		namespace.Controller.after.should.be.a('function');
+		namespace.Controller.skipAfter.should.be.a('function');
+	});
+	it('Reflection', () => {
+		should.exist(namespace.Reflection);
+		should.exist(namespace.Reflection.getFunctionParams);
+		should.exist(namespace.Reflection.isAnnotated);
+		should.exist(namespace.Reflection.annotate);
+		should.exist(namespace.Reflection.invoke);
+		should.exist(namespace.Reflection.invokeSuper);
+		should.exist(namespace.Reflection.construct);
+		should.exist(namespace.Reflection.symbols);
+		namespace.Reflection.getFunctionParams.should.be.a('function');
+		namespace.Reflection.isAnnotated.should.be.a('function');
+		namespace.Reflection.annotate.should.be.a('function');
+		namespace.Reflection.invoke.should.be.a('function');
+		namespace.Reflection.invokeSuper.should.be.a('function');
+		namespace.Reflection.construct.should.be.a('function');
 	});
 });